+++ R/caching.R
@@ -3,10 +3,10 @@
 #' If the target file exists it will be loaded unless
 #' the 'overwrite' argument is set.
 #'
-#' @param file Where to cache the R object.
-#' @param object An R object.
-#' @param overwrite Whether to overwrite the cached object.
-#' @param ... Additional arguments for loading from cache.
+#' @param file Where to cache the R object
+#' @param object An R object
+#' @param overwrite Whether to overwrite the cached object
+#' @param ... Unused, will warn
 #'
 #' @return The fresh/cached object.
 #' @export
@@ -19,23 +19,44 @@
 #' unlink(target)
 #' }
 with_cache <- function(file, object, overwrite = FALSE, ...) {
-  cache_fns <- get_cache_fns_from_ext(file)
-  if (isFALSE(file.exists(file)) || isTRUE(overwrite)) {
+
+  if (length(list(...)) > 0) {
+    warning("with_cache() no longer uses arguments provided via `...`")
+  }
+
+  ext <- tolower(tools::file_ext(file))
+  cache_fns <- switch(ext, fst = fst_cache_functions(), rds_cache_functions())
+
+  if (isTRUE(cache_is_invalid(file)) || isTRUE(overwrite)) {
     message("Caching object to ", file)
     cache_fns$write(object, file)
   }
+
   # Always load from cache as a check against failure
   message("Loading object from ", file)
-  cache_fns$read(file, ...)
+  cache_fns$read(file)
 }
 
 
-# Determine which reading/writing functions to use based on input file
-get_cache_fns_from_ext <- function(path) {
-  ext <- tolower(tools::file_ext(path))
-  if (ext == "fst") {
-    check_for_packages("fst")
-    return(list(read = fst::read_fst, write = fst::write_fst))
-  }
+cache_is_invalid <- function(file) {
+  isFALSE(file.exists(file))
+}
+
+
+rds_cache_functions = function() {
   list(read = readRDS, write = saveRDS)
+}
+
+
+fst_cache_functions = function() {
+  useDT = (
+    getOption("tt.cache.prefer_data_table", FALSE) &&
+    requireNamespace("data.table", quietly = TRUE)
+  )
+
+  # Use the same kwargs as readRDS/saveRDS
+  list(
+    read = function(file) fst::read_fst(path = file, as.data.table = useDT),
+    write = function(object, file) fst::write_fst(object, file)
+  )
 }